Añadida sección de indentación

+++ codigo/4-variables/input_3.py
@@ -1,3 +1,3 @@
-x = int(input("Ingrese un número: "))
+x = float(input("Ingrese un número: "))
 
 print(x + 10)
